basic test of mezzanine withdrawal
test needs to be more precise (needs to check that withdrawn values are
as expected), but this is what I have in place right now.

bug fix: was using the wrong loop counter variable in mezzanine
entitlement calculations!!!

bug fix: was using the wrong values to decrement
`unallocatedEntitlements` as the entitlement caclulation process was
marching on.

also realized that there never was any overflow issue, that all
anomalies I was seeing were due to the aforementioned bugs, so
converted those comparisons to tranche array lenghts back to their more
readable original forms.